Extract start and stop, with new stop behavior
