+++ packages/flutter_tools/lib/src/linux/linux_device.dart
@@ -2,17 +2,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import '../application_package.dart';
-import '../base/io.dart';
 import '../base/platform.dart';
-import '../base/process_manager.dart';
 import '../build_info.dart';
-import '../desktop.dart';
 import '../desktop_device.dart';
 import '../device.dart';
-import '../globals.dart';
 import '../project.dart';
-import '../protocol_discovery.dart';
 import 'application_package.dart';
 import 'build_linux.dart';
 import 'linux_workflow.dart';
@@ -26,61 +20,10 @@ class LinuxDevice extends DesktopDevice {
   );
 
   @override
-  void clearLogs() { }
-
-  @override
-  DeviceLogReader getLogReader({ ApplicationPackage app }) {
-    return _logReader;
-  }
-  final DesktopLogReader _logReader = DesktopLogReader();
-
-  @override
   bool isSupported() => true;
 
   @override
   String get name => 'Linux';
-
-  @override
-  Future<LaunchResult> startApp(
-    covariant LinuxApp package, {
-    String mainPath,
-    String route,
-    DebuggingOptions debuggingOptions,
-    Map<String, dynamic> platformArgs,
-    bool prebuiltApplication = false,
-    bool ipv6 = false,
-  }) async {
-    _lastBuiltMode = debuggingOptions.buildInfo.mode;
-    if (!prebuiltApplication) {
-      await buildLinux(
-        FlutterProject.current().linux,
-        debuggingOptions.buildInfo,
-        target: mainPath,
-      );
-    }
-    final Process process = await processManager.start(<String>[
-      package.executable(debuggingOptions?.buildInfo?.mode)
-    ]);
-    if (debuggingOptions?.buildInfo?.isRelease == true) {
-      return LaunchResult.succeeded();
-    }
-    _logReader.initializeProcess(process);
-    final ProtocolDiscovery observatoryDiscovery = ProtocolDiscovery.observatory(_logReader);
-    try {
-      final Uri observatoryUri = await observatoryDiscovery.uri;
-      return LaunchResult.succeeded(observatoryUri: observatoryUri);
-    } catch (error) {
-      printError('Error waiting for a debug connection: $error');
-      return LaunchResult.failed();
-    } finally {
-      await observatoryDiscovery.cancel();
-    }
-  }
-
-  @override
-  Future<bool> stopApp(covariant LinuxApp app) async {
-    return killProcess(app.executable(_lastBuiltMode));
-  }
 
   @override
   Future<TargetPlatform> get targetPlatform async => TargetPlatform.linux_x64;
@@ -90,8 +33,23 @@ class LinuxDevice extends DesktopDevice {
     return flutterProject.linux.existsSync();
   }
 
-  // Track the last built mode from startApp.
-  BuildMode _lastBuiltMode;
+  @override
+  Future<void> buildForDevice(
+    covariant LinuxApp package, {
+    String mainPath,
+    BuildInfo buildInfo,
+  }) async {
+    await buildLinux(
+      FlutterProject.current().linux,
+      buildInfo,
+      target: mainPath,
+    );
+  }
+
+  @override
+  String executablePathForDevice(covariant LinuxApp package, BuildMode buildMode) {
+    return package.executable(buildMode);
+  }
 }
 
 class LinuxDevices extends PollingDeviceDiscovery {
