+++ packages/flutter_tools/lib/src/windows/windows_device.dart
@@ -4,17 +4,13 @@
 
 import 'package:meta/meta.dart';
 
-import '../application_package.dart';
 import '../base/io.dart';
 import '../base/platform.dart';
 import '../base/process.dart';
 import '../build_info.dart';
-import '../desktop.dart';
 import '../desktop_device.dart';
 import '../device.dart';
-import '../globals.dart';
 import '../project.dart';
-import '../protocol_discovery.dart';
 import 'application_package.dart';
 import 'build_windows.dart';
 import 'windows_workflow.dart';
@@ -28,68 +24,10 @@ class WindowsDevice extends DesktopDevice {
   );
 
   @override
-  void clearLogs() { }
-
-  @override
-  DeviceLogReader getLogReader({ ApplicationPackage app }) {
-    return _logReader;
-  }
-  final DesktopLogReader _logReader = DesktopLogReader();
-
-  @override
   bool isSupported() => true;
 
   @override
   String get name => 'Windows';
-
-  @override
-  Future<LaunchResult> startApp(
-    covariant WindowsApp package, {
-    String mainPath,
-    String route,
-    DebuggingOptions debuggingOptions,
-    Map<String, dynamic> platformArgs,
-    bool prebuiltApplication = false,
-    bool ipv6 = false,
-  }) async {
-    if (!prebuiltApplication) {
-      await buildWindows(
-        FlutterProject.current().windows,
-        debuggingOptions.buildInfo,
-        target: mainPath,
-      );
-    }
-    final Process process = await processUtils.start(<String>[
-      package.executable(debuggingOptions?.buildInfo?.mode)
-    ]);
-    if (debuggingOptions?.buildInfo?.isRelease == true) {
-      return LaunchResult.succeeded();
-    }
-    _logReader.initializeProcess(process);
-    final ProtocolDiscovery observatoryDiscovery = ProtocolDiscovery.observatory(_logReader);
-    try {
-      final Uri observatoryUri = await observatoryDiscovery.uri;
-      return LaunchResult.succeeded(observatoryUri: observatoryUri);
-    } catch (error) {
-      printError('Error waiting for a debug connection: $error');
-      return LaunchResult.failed();
-    } finally {
-      await observatoryDiscovery.cancel();
-    }
-  }
-
-  @override
-  Future<bool> stopApp(covariant WindowsApp app) async {
-    // Assume debug for now.
-    final List<String> process = runningProcess(app.executable(BuildMode.debug));
-    if (process == null) {
-      return false;
-    }
-    final RunResult result = await processUtils.run(
-      <String>['Taskkill', '/PID', process.first, '/F'],
-    );
-    return result.exitCode == 0;
-  }
 
   @override
   Future<TargetPlatform> get targetPlatform async => TargetPlatform.windows_x64;
@@ -97,6 +35,24 @@ class WindowsDevice extends DesktopDevice {
   @override
   bool isSupportedForProject(FlutterProject flutterProject) {
     return flutterProject.windows.existsSync();
+  }
+
+  @override
+  Future<void> buildForDevice(
+    covariant WindowsApp package, {
+    String mainPath,
+    BuildInfo buildInfo,
+  }) async {
+    await buildWindows(
+      FlutterProject.current().windows,
+      buildInfo,
+      target: mainPath,
+    );
+  }
+
+  @override
+  String executablePathForDevice(covariant WindowsApp package, BuildMode buildMode) {
+    return package.executable(buildMode);
   }
 }
